+++ package.json
@@ -104,6 +104,7 @@
     "eslint-plugin-prettier": "3.1.2",
     "eslint-plugin-promise": "4.2.1",
     "eslint-plugin-react": "7.19.0",
+    "eslint-plugin-react-hooks": "3.0.0",
     "fast-xml-parser": "3.16.0",
     "husky": "4.2.3",
     "jest": "25.2.6",
